Store long value in CharacterObject
to avoid unnecessary int<->long conversions

+++ src/de.hpi.swa.graal.squeak/src/de/hpi/swa/graal/squeak/image/SqueakImageChunk.java
@@ -215,7 +215,7 @@ public final class SqueakImageChunk {
             case SqueakImageConstants.SMALL_INTEGER_TAG: // SmallInteger
                 return ptr >> SqueakImageConstants.NUM_TAG_BITS;
             case SqueakImageConstants.CHARACTER_TAG: // Character
-                return CharacterObject.valueOf((int) (ptr >> SqueakImageConstants.NUM_TAG_BITS));
+                return CharacterObject.valueOf(ptr >> SqueakImageConstants.NUM_TAG_BITS);
             case SqueakImageConstants.SMALL_FLOAT_TAG:
                 /* SmallFloat (see Spur64BitMemoryManager>>#smallFloatBitsOf:). */
                 long valueWithoutTag = ptr >>> SqueakImageConstants.NUM_TAG_BITS;
